requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
qrcode[pil]==7.4.2
itsdangerous==2.2.0
Jinja2==3.1.4
Pillow==10.4.0

